combine types.go and unversioned.go
